Move lca component out of asmt template folder
