+++ src/ggrc-client/js/components/assessment-template-attributes/add-template-field/add-template-field.stache
