+++ src/ggrc-client/js/components/assessment-template-attributes/template-field/template-field.js
