+++ src/ggrc-client/js/components/assessment-template-attributes/templates/assessment-template-attributes.stache
@@ -1,0 +1,55 @@
+{{!
+    Copyright (C) 2019 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+}}
+
+<div class="attr-titles">
+  <ul>
+    <li class="row-fluid">
+      <div class="span-custom2-and-half">
+        <h6>Attribute Title</h6>
+      </div>
+      <div class="span-custom1-and-half">
+        <h6>Attribute type</h6>
+      </div>
+      <div class="span2">
+        <h6>Attribute values</h6>
+      </div>
+      <div class="span1 centered">
+        <h6>Evidence File Required</h6>
+      </div>
+      <div class="span1 centered">
+        <h6>Evidence URL Required</h6>
+      </div>
+      <div class="span1 centered">
+        <h6>Comment required</h6>
+      </div>
+      <div class="span1 centered">
+        <h6>Mandatory</h6>
+      </div>
+      <div class="span2">
+        &nbsp;
+      </div>
+    </li>
+  </ul>
+</div>
+<div class="sortable-list new-attr-list">
+  <ul>
+    <li class="row-fluid add-field">
+      <add-template-field
+        types:from="types"
+        fields:from="fields">
+      </add-template-field>
+    </li>
+    {{#fields}}
+      {{^if _pending_delete}}
+          <li class="row-fluid sortable-item" {{canData 'field'}}>
+              <template-field
+                field:from="{.}"
+                types:from="types"
+                on:remove="fieldRemoved"></template-field>
+          </li>
+      {{/if}}
+    {{/fields}}
+  </ul>
+</div>
