+++ src/ggrc-client/js/controllers/modals/modals_controller.js
@@ -18,10 +18,8 @@ import '../../components/issue-tracker/issue-tracker-switcher';
 import '../../components/access-control-list/access-control-list-roles-helper';
 import '../../components/assessment/assessment-people';
 import '../../components/assessment/assessment-object-type-dropdown';
-import '../../components/assessment-templates/assessment-template-attributes/assessment-template-attributes';
+import '../../components/assessment-template-attributes/assessment-template-attributes';
 import '../../components/assessment-templates/people-list/people-list';
-import '../../components/assessment-templates/template-field/template-field';
-import '../../components/assessment-templates/add-template-field/add-template-field';
 import '../../components/textarea-array/textarea-array';
 import '../../components/object-list-item/object-list-item-updater';
 import '../../components/related-objects/related-documents';
