+++ src/ggrc-client/js/templates/assessment_templates/modal_content.stache
@@ -154,56 +154,6 @@
           <div class="row-fluid">
             <div class="custom-attr-list span12">
               <assessment-template-attributes fields:from="instance.custom_attribute_definitions">
-                <div class="attr-titles">
-                  <ul>
-                    <li class="row-fluid">
-                      <div class="span-custom2-and-half">
-                        <h6>Attribute Title</h6>
-                      </div>
-                      <div class="span-custom1-and-half">
-                        <h6>Attribute type</h6>
-                      </div>
-                      <div class="span2">
-                        <h6>Attribute values</h6>
-                      </div>
-                      <div class="span1 centered">
-                        <h6>Evidence File Required</h6>
-                      </div>
-                      <div class="span1 centered">
-                        <h6>Evidence URL Required</h6>
-                      </div>
-                      <div class="span1 centered">
-                        <h6>Comment required</h6>
-                      </div>
-                      <div class="span1 centered">
-                        <h6>Mandatory</h6>
-                      </div>
-                      <div class="span2">
-                        &nbsp;
-                      </div>
-                    </li>
-                  </ul>
-                </div>
-                <div class="sortable-list new-attr-list">
-                  <ul>
-                    <li class="row-fluid add-field">
-                      <add-template-field
-                        types:from="types"
-                        fields:from="fields">
-                      </add-template-field>
-                    </li>
-                    {{#fields}}
-                      {{^if _pending_delete}}
-                          <li class="row-fluid sortable-item" {{canData 'field'}}>
-                              <template-field
-                                field:from="{.}"
-                                types:from="types"
-                                on:remove="fieldRemoved"></template-field>
-                          </li>
-                      {{/if}}
-                    {{/fields}}
-                  </ul>
-                </div>
               </assessment-template-attributes>
             </div>
           </div>
